Detect unsupported system by parsing stderr :bug: :sparkles:
Sometimes, particularly when using import-from-derivation, a builder for
an unsupported system is needed during evaluation - therefore an error
occurs before we have checked if the system is supported. This can be
remedied by parsing stderr for messages about unsupported systems when
an error occurs, rather than trying to predict the failure.

Additionally, by only checking the system when an error occurs, we gain
the ability to have successful builds on an unsupported system if the
output is already available from a binary cache.